Remove the actorInstitution
The actorInstitution could be removed because the institution switcher
could be removed, because it is no longer useful after the FGA changes.
This because earlier all overviews were from the view point of an
institution but now it's from a registration authority.

These changes are needed to reflect the changes in RA.