Add the SelfVet command
This command will trigger the self vetting event on the middleware

In adition to the regular data, an authoring second factor id (of the
user itself) must be passed. This id will be used to vet the verified
token with. It must have an equal or higher loa then the verified one.